READ ME
In this example, we're trying to find the minimum value of the function f(x) = x[0]**2 + x[1]**2 using PSO. PSO is a population-based optimization algorithm, so we start by defining the PSO parameters: the number of particles (num_particles), the maximum number of iterations (max_iterations), the cognitive (c1) and social (c2) parameters, and the inertia weight (w). We also define the search space (bounds), which is the range of possible values for each dimension of the input vector.

Next, we define the Particle class. Each particle has a position, velocity, best position, best fitness, and current fitness. The update_position, update_velocity, and evaluate_fitness methods are used to update the particle's position and velocity, and to evaluate its fitness using the problem function.

The PSO algorithm itself is implemented in the pso function. We start by initializing a population of num_particles particles, and we keep track of the global best position and fitness. In each iteration, we evaluate the fitness of each particle, update the global best position and fitness.